Included Analog Sensors in DTC system
Had to update the ADC "eGetUpdate()" method to return a struct called "ADC_Result" that contains the value and SPI error code for the respect sensor. Any calls to this method need to take this into account, likely just by adding a ".value" to the end of the call to ensure proper usage.